src/button_control/python/requirements.txt
RPi.GPIO>=0.6.4